+++ coffee-cdi/src/main/java/hu/icellmobilsoft/coffee/cdi/annotation/Version.java
@@ -17,7 +17,7 @@
  * limitations under the License.
  * #L%
  */
-package hu.icellmobilsoft.coffee.dto.annotation;
+package hu.icellmobilsoft.coffee.cdi.annotation;
 
 import static java.lang.annotation.ElementType.TYPE;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
