+++ coffee-rest/src/main/java/hu/icellmobilsoft/coffee/rest/validation/xml/XmlMessageBodyReaderBase.java
@@ -33,8 +33,8 @@ import jakarta.ws.rs.ext.MessageBodyReader;
 import org.apache.commons.lang3.ArrayUtils;
 
 import hu.icellmobilsoft.coffee.dto.exception.BaseException;
-import hu.icellmobilsoft.coffee.dto.annotation.validation.xml.ValidateXML;
-import hu.icellmobilsoft.coffee.dto.annotation.validation.xml.ValidateXMLs;
+import hu.icellmobilsoft.coffee.cdi.annotation.xml.ValidateXML;
+import hu.icellmobilsoft.coffee.cdi.annotation.xml.ValidateXMLs;
 import hu.icellmobilsoft.coffee.rest.validation.xml.exception.BaseProcessingExceptionWrapper;
 
 /**
