+++ coffee-tool/src/test/java/hu/icellmobilsoft/coffee/tool/utils/annotation/RangeUtilTest.java
@@ -29,8 +29,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-import hu.icellmobilsoft.coffee.dto.annotation.Range;
-import hu.icellmobilsoft.coffee.dto.annotation.Range.RangeLiteral;
+import hu.icellmobilsoft.coffee.cdi.annotation.Range;
+import hu.icellmobilsoft.coffee.cdi.annotation.Range.RangeLiteral;
 
 /**
  * RangeUtil test class
